pyproject.toml
[project]
name = "report-draft-beta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.5",
    "streamlit>=1.46.1",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pypandoc>=1.15",
    "pypandoc-binary>=1.15",
    "python-docx>=1.2.0",
]
